Fix mypy errors

+++ test/python/compiler/naive_local_resynth/test_local_resynth.py
@@ -31,7 +31,11 @@ class TestNaiveLocResynthOptPass(TestCase):
             R(self.circuit, "R", 0, [0, 1, -np.pi / 5, np.pi / 7], self.circuit.dimensions[0]),
             R(self.circuit, "R", 0, [1, 2, -np.pi, np.pi / 2], self.circuit.dimensions[0]),
             CustomMulti(
-                self.circuit, "CUm", [0, 1, 2], np.identity(27), [self.circuit.dimensions[i] for i in [0, 1, 2]]
+                self.circuit,
+                "CUm",
+                [0, 1, 2],
+                np.identity(27, dtype=np.complex128),
+                [self.circuit.dimensions[i] for i in [0, 1, 2]],
             ),
             R(self.circuit, "R", 2, [0, 1, np.pi, np.pi / 2], self.circuit.dimensions[2]),
             CEx(self.circuit, "CEx", [0, 1], None, [self.circuit.dimensions[i] for i in [0, 1]]),
